chore: update to tsx
